Refactor metrics to be outside of the app object
