+++ apps/glusterfs/app_topology.go
@@ -1,0 +1,110 @@
+//
+// Copyright (c) 2015 The heketi Authors
+//
+// This file is licensed to you under your choice of the GNU Lesser
+// General Public License, version 3 or any later version (LGPLv3 or
+// later), or the GNU General Public License, version 2 (GPLv2), in all
+// cases as published by the Free Software Foundation.
+//
+
+package glusterfs
+
+import (
+	"github.com/boltdb/bolt"
+	"github.com/heketi/heketi/pkg/glusterfs/api"
+)
+
+func (a *App) TopologyInfo() (*api.TopologyInfoResponse, error) {
+	topo := &api.TopologyInfoResponse{
+		ClusterList: make([]api.Cluster, 0),
+	}
+
+	err := a.db.View(func(tx *bolt.Tx) error {
+		clusters, err := ClusterList(tx)
+		if err != nil {
+			return err
+		}
+
+		for _, cluster := range clusters {
+			clusteri, err := clusterInfo(tx, cluster)
+			if err != nil {
+				return err
+			}
+			cluster := api.Cluster{
+				Id:      clusteri.Id,
+				Volumes: make([]api.VolumeInfoResponse, 0),
+				Nodes:   make([]api.NodeInfoResponse, 0),
+				ClusterFlags: api.ClusterFlags{
+					Block: clusteri.Block,
+					File:  clusteri.File,
+				},
+			}
+			cluster.Id = clusteri.Id
+
+			for _, volumes := range clusteri.Volumes {
+				volumesi, err := volumeInfo(tx, volumes)
+				if err != nil {
+					return err
+				}
+				if volumesi.Cluster == cluster.Id {
+					cluster.Volumes = append(cluster.Volumes, *volumesi)
+				}
+			}
+
+			for _, node := range clusteri.Nodes {
+				nodei, err := nodeInfo(tx, string(node))
+				if err != nil {
+					return err
+				}
+				cluster.Nodes = append(cluster.Nodes, *nodei)
+			}
+			topo.ClusterList = append(topo.ClusterList, cluster)
+		}
+		return nil
+	})
+
+	return topo, err
+}
+
+func clusterInfo(tx *bolt.Tx, id string) (*api.ClusterInfoResponse, error) {
+	var info *api.ClusterInfoResponse
+	entry, err := NewClusterEntryFromId(tx, id)
+	if err != nil {
+		return info, err
+	}
+
+	info, err = entry.NewClusterInfoResponse(tx)
+	if err != nil {
+		return info, err
+	}
+
+	err = UpdateClusterInfoComplete(tx, info)
+
+	return info, err
+}
+
+func volumeInfo(tx *bolt.Tx, id string) (*api.VolumeInfoResponse, error) {
+	var info *api.VolumeInfoResponse
+
+	entry, err := NewVolumeEntryFromId(tx, id)
+	if err != nil {
+		return info, err
+	}
+
+	info, err = entry.NewInfoResponse(tx)
+
+	return info, err
+}
+
+func nodeInfo(tx *bolt.Tx, id string) (*api.NodeInfoResponse, error) {
+	var info *api.NodeInfoResponse
+
+	entry, err := NewNodeEntryFromId(tx, id)
+	if err != nil {
+		return info, err
+	}
+
+	info, err = entry.NewInfoReponse(tx)
+
+	return info, err
+}
