+++ main.go
@@ -21,6 +21,7 @@ import (
 	"github.com/heketi/heketi/apps"
 	"github.com/heketi/heketi/apps/glusterfs"
 	"github.com/heketi/heketi/middleware"
+	"github.com/heketi/heketi/pkg/metrics"
 	"github.com/spf13/cobra"
 	"github.com/urfave/negroni"
 
@@ -223,7 +224,7 @@ func main() {
 			fmt.Fprint(w, "Hello from Heketi")
 		})
 
-	router.Methods("GET").Path("/metrics").Name("Metrics").HandlerFunc(app.Metrics())
+	router.Methods("GET").Path("/metrics").Name("Metrics").HandlerFunc(metrics.NewMetricsHandler(app))
 
 	// Create a router and do not allow any routes
 	// unless defined.
